+++ lib/config/app_config.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 class AppConfig {
   static int flashcardTextMaxLength = 200;
-  static int maxFlashcardInGroup = 100;
+  static int maxFlashcardInGroup = 200;
   static int quickReviewQuestionNumber = 10;
+  static int maxExamStepsToShowStepIndicator = 25;
 }